requirements.txt
requests~=2.31.0
m3u8~=3.6.0
yarl~=1.9.2
regex~=2023.10.3
beautifulsoup4~=4.12.2
js2xml~=0.5.0
xmltodict~=0.13.0
pycryptodome~=3.19.0
